Add ens names to query
